pyproject.toml
[project]
name = "app-routes"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dependency-injector>=4.46.0",
    "fastapi>=0.115.11",
    "pydantic[email]>=2.10.6",
    "pydantic-settings>=2.8.1",
    "python-dotenv>=1.0.1",
    "sqlalchemy>=2.0.38",
    "uvicorn>=0.34.0",
    "pyjwt>=2.10.1",
    "bcrypt>=4.3.0",
    "asyncpg>=0.30.0",
    "alembic>=1.15.1",
    "cryptography>=44.0.2",
    "pytest-asyncio>=0.25.3",
]

[tool.ruff]
line-length = 100  
target-version = "py311"

[tool.ruff.lint]
select = ["ALL"]  
ignore = [
    "E501", "D", "UP035", "ANN401",
    "B010", "B009", "SIM117", "EM101",
    "TRY003", "FAST002", "B008", "TID252"
]  


[tool.pytest.ini_options]
pythonpath = ["src", "tests"]
testpaths = ["tests"]
log_cli = true
log_cli_level = "INFO"

